fix bug

+++ MastactvaBase/names.h
@@ -208,7 +208,7 @@ Q_DECL_UNUSED static const char * g_mergeIdFieldName = "mergeid";
 
 
 // drawing data
-Q_DECL_UNUSED static const int g_renderTeps = 1e-5;
+Q_DECL_UNUSED static const double g_renderTeps = 1e-5;
 
 
 #endif // NAMES_H
